+++ esp-wifi/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bump MSRV to 1.83 (#2615)
+- Bump smoltcp to 0.12.0 (#2849)
 
 ### Fixed